requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
datetime
dill
pyyaml
-e .
## This -e . is used to refer the setup.py
## also useful when we deploy in cloud 